+++ apps/web/src/app/[locale]/(space)/page.tsx
@@ -18,8 +18,8 @@ import {
 } from "@/app/components/page-layout";
 import { Trans } from "@/components/trans";
 import { IfCloudHosted } from "@/contexts/environment";
-import { loadUpcomingEventsCount } from "@/data/event";
 import { loadActiveSpace } from "@/data/space";
+import { loadUpcomingEventsCount } from "@/features/scheduled-event/data";
 import { getTranslation } from "@/i18n/server";
 import { FeedbackAlert } from "./feedback-alert";
